feat: introduces LEMACS custom option: lemacs_lsp_client
You can now choose if you want 'lsp-mode or 'eglot to be your
lsp client.

Eglot breaks TUI/GUI experience, since it depends on eldoc-box on GUI,
which is not available on TUI.

This is the first step to toggle options in order to offer the user
a way to try newer things if the user wants to.